Refactor: spawn out sbd-watchdog.c
